refactor: Add live-preview mode
When build with "LOCAL_PREVIEW" or "VERCEL_ENV=preview", the project is
builded in live preview mode. This change allows for conditional logic
in the algolia-generate.ts file, preventing the execution of certain
code when in live preview mode. Also in live preview mode tracking tags
( umami and google analytics ) won't be added.